1、解决本机int转型失败问题 2、增加按照序号排序的功能 3、完成数据生成配置界面功能

+++ DBDataGenerator/DataModels/ColumnSchema.cs
@@ -24,7 +24,7 @@ namespace DBDataGenerator.DataModels
         /// <summary>
         /// 序号
         /// </summary>
-        public UInt64? ORDINAL_POSITION { get; set; } = 0;
+        public UInt32? ORDINAL_POSITION { get; set; } = 0;
 
 
         /// <summary>
